pyproject.toml
[project]
name = "mcap-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "av>=15.0.0",
    "foxglove-sdk>=0.10.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
]
